Adds ability to use signed embeds and pass the 'expires' parameter wer when calling player.js. This prevents local modification of embed codes and allows for locking videos by setting to either 'Paused' or with a 'Flight Time' in the past in backlot. Resolves #69
This commit probably breaks thickbox support in the ooayala module but since we're not using that on Drupalize.me it should be fine for now and I'll work on cleaning that up at another time.

We'll also need to decide on a strategy regarding using 'Paused' or 'Flight Time' for videos in backlot. Refs #105